IPv6 support (#3952)
Implemented support for IPv6 connections. The server now listens on an
IPv6 socket unless it's configured to only bind to a specific IPv4
address. The address is now stored as a varchar in the database.
The internal representation of the address has been changed to the
asio-native one for simplicity.

- Remove convertIPToString()

This function is not used anymore.

- Switch to ip column to VARBINARY(16)

varbinary(16) offers greater space efficiency than a character string.
Changed the bound acceptor address to a static to prevent constructing
it on every connection

Signed-off-by: Ranieri Althoff <ranisalt@gmail.com>
Co-authored-by: Damian Jarek <damian.jarek93@gmail.com>
Co-authored-by: Evil Puncker <EPuncker@users.noreply.github.com>