Sync with TFS (#93)
* fix variable scope in npc.lua (#4187)

* Add possibility to cancel LookDescription by returning an empty string (#4176)

* IPv6 support (#3952)

Implemented support for IPv6 connections. The server now listens on an
IPv6 socket unless it's configured to only bind to a specific IPv4
address. The address is now stored as a varchar in the database.
The internal representation of the address has been changed to the
asio-native one for simplicity.

- Remove convertIPToString()

This function is not used anymore.

- Switch to ip column to VARBINARY(16)

varbinary(16) offers greater space efficiency than a character string.
Changed the bound acceptor address to a static to prevent constructing
it on every connection

Signed-off-by: Ranieri Althoff <ranisalt@gmail.com>
Co-authored-by: Damian Jarek <damian.jarek93@gmail.com>
Co-authored-by: Evil Puncker <EPuncker@users.noreply.github.com>

* fix io.write not writing to console (#4209)

* Logging code improvements (#4218)

* fix io.write not writing to console

* logging code improvements

Co-authored-by: Evil Puncker <EPuncker@users.noreply.github.com>

* Update schema.sql

Signed-off-by: Ranieri Althoff <ranisalt@gmail.com>
Co-authored-by: Łukasz Zbiżek <Zbizu@users.noreply.github.com>
Co-authored-by: Sarah Wesker <sarahelizabetwesker@gmail.com>
Co-authored-by: Ranieri Althoff <1993083+ranisalt@users.noreply.github.com>
Co-authored-by: Damian Jarek <damian.jarek93@gmail.com>
Co-authored-by: Evil Puncker <EPuncker@users.noreply.github.com>